homectl/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "homectl"
version = "0.1.0"
dependencies = ["typer", "paho-mqtt"]

[project.scripts]
homectl = "homectl:app"

[tool.setuptools]
py-modules = ["homectl"]
